ImageCompression/requirements.txt
pillow
scipy
numpy
pyqt5
prettytable
pyssim
pandas